backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1

# Auth
pyjwt[crypto]==2.10.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0

# Validation & settings
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Storage
boto3==1.36.2
boto3-stubs[s3]==1.36.2

# Background jobs
arq==0.26.1
redis==5.2.1

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# HTTP client
httpx==0.28.1

# Utilities
python-slugify==8.0.4
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Dev
ruff==0.8.6
